Create conductor.py
This is Diane's version of the conductor / brain code. It has some infrastructure and a lot of testing abilities built into it, and still has a lot left to code. It seems to be pretty orthogonal to Jeff's brain code. I believe that Jeff offered to merge these two into one working program.